+++ docker-compose.yml
@@ -13,9 +13,8 @@ services:
       - STORAGE_API_TOKEN
       - STORAGE_API_LINKING_TOKEN
       - STORAGE_API_MAINTENANCE_URL
-      - STORAGE_API_PROJECT_IDS_IN_ORGANIZATION
-      - STORAGE_API_PROJECT_IDS_NOT_IN_ORGANIZATION
-      - STORAGE_API_PROJECT_ID_NOT_AVAILABLE_TO_LINK
+      - STORAGE_API_TOKEN_IN_SAME_ORGANIZATION
+      - STORAGE_API_TOKEN_IN_OTHER_ORGANIZATION
       - REDSHIFT_HOSTNAME
       - REDSHIFT_USER
       - REDSHIFT_PASSWORD
