+++ set-env.sharing.template.sh
@@ -1,7 +1,6 @@
 export STORAGE_API_URL=http://connection-apache/
 export STORAGE_API_TOKEN=
 export STORAGE_API_LINKING_TOKEN=
+export STORAGE_API_TOKEN_IN_SAME_ORGANIZATION=
+export STORAGE_API_TOKEN_IN_OTHER_ORGANIZATION=
 export STORAGE_API_MAINTENANCE_URL=https://maintenance-testing.us-east-1.keboola.com/
-export STORAGE_API_PROJECT_IDS_IN_ORGANIZATION=
-export STORAGE_API_PROJECT_ID_NOT_AVAILABLE_TO_LINK=
-export STORAGE_API_PROJECT_IDS_NOT_IN_ORGANIZATION=
