+++ tests/bootstrap.php
@@ -6,9 +6,8 @@ define('STORAGE_API_URL', getenv('STORAGE_API_URL'));
 define('STORAGE_API_TOKEN', getenv('STORAGE_API_TOKEN'));
 define('STORAGE_API_LINKING_TOKEN', getenv('STORAGE_API_LINKING_TOKEN'));
 define('STORAGE_API_MAINTENANCE_URL', getenv('STORAGE_API_MAINTENANCE_URL'));
-define('STORAGE_API_PROJECT_IDS_IN_ORGANIZATION', getenv('STORAGE_API_PROJECT_IDS_IN_ORGANIZATION'));
-define('STORAGE_API_PROJECT_IDS_NOT_IN_ORGANIZATION', getenv('STORAGE_API_PROJECT_IDS_NOT_IN_ORGANIZATION'));
-define('STORAGE_API_PROJECT_ID_NOT_AVAILABLE_TO_LINK', getenv('STORAGE_API_PROJECT_ID_NOT_AVAILABLE_TO_LINK'));
+define('STORAGE_API_TOKEN_IN_SAME_ORGANIZATION', getenv('STORAGE_API_TOKEN_IN_SAME_ORGANIZATION'));
+define('STORAGE_API_TOKEN_IN_OTHER_ORGANIZATION', getenv('STORAGE_API_TOKEN_IN_OTHER_ORGANIZATION'));
 define('REDSHIFT_HOSTNAME', getenv('REDSHIFT_HOSTNAME'));
 define('REDSHIFT_USER', getenv('REDSHIFT_USER'));
 define('REDSHIFT_PASSWORD', getenv('REDSHIFT_PASSWORD'));
